requirements.txt
requests
beautifulsoup4
PyPDF2
pdfplumber
pandas
python-dotenv
python-telegram-bot